Split down the application to the different layers (Controller, Service, Repository
