+++ src/main/java/com/example/fizzbuzz/model/FizzBuzz.java
@@ -1,10 +1,8 @@
 package com.example.fizzbuzz.model;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
